tranform management section lazy loading (#62408)
